backend/requirements.txt
appdata==2.1.2
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.0.3
colored==1.4.3
Django==4.2.5
django-parler==2.3
django-rosetta==0.9.9
djangorestframework==3.14.0
idna==3.4
polib==1.2.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
to-requirements.txt==1.3.0
typing_extensions==4.8.0
urllib3==2.0.5
django-rest-framework-i18n==1.0.1


